Updated serial code for part 6 of lab
Updated the serial code for part 6 of lab 2. This stand alone code is
now appropriate for all lab parts.

If you have not received serial in the last second then the onboard LED
(pin13) should be off.  The light blink interval will set based on the
switch/button/pot as per the lab manual.

If you are receiving serial then it will turn on the onbaord LED (pin13)
and it will use the received value for the light blink interval.

You can test the serial by jumping pins D0 and D1 together. The
microcontroller will essentially talk to itself. LED13 should turn on.